+++ lib/langtool_web/views/session_view.ex
@@ -1,0 +1,3 @@
+defmodule LangtoolWeb.SessionView do
+  use LangtoolWeb, :view
+end